Task 6: Minor tweaks
Now if a node "does not exist" because it isn't reachable (wall etc.)
the edge towards it is set to -1 rather than 0 , the reason being that
the source e.g. is initialized with 0 distance. (It's just a tweak in
the concept not really semantic)
    When a door is closed it should not be set to -1 (old 0) because it
is still openable, rather it should be set to 2 beacause passing it needs
one more step